mmap, updates for wsketch, add support for 32-bit integer weights. (#29)
This commit adds mmap support for the signature matrix, which can substantially reduce peak memory footprint.

Further, the wsketch command has been extended and simplified; large arrays are now mmap'd for CSR format, and in the future we plan to update for simpler modes. Additionally, we added support for 32-bit integer weights.